requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
GitPython==3.1.43
psycopg2-binary==2.9.10
tzdata==2024.2